+++ src/schemas/validation.rs
@@ -23,7 +23,6 @@ pub struct SchemaValidationContext {
   _schema: Schema,
 }
 
-
 impl SchemaValidationContext {
   /// Create a schema validation context from a parser object
   pub fn from_parser(parser: &mut SchemaParserContext) -> Result<Self, Vec<StructuredError>> {